backend/main_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==3.2.2
python-jose
pydantic<2
psycopg2
python-multipart

torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
ultralytics==8.3.239
timm==1.0.22
pillow

imageio-ffmpeg

python-dotenv
psycopg2-binary
httpx
pandas
lightgbm
scikit-learn
Faker==19.3.0

edge-tts